backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
sqlalchemy==2.0.23
python-multipart==0.0.6
aiofiles==23.1.0
python-dotenv==1.0.1
pyjwt==2.8.1
requests==2.31.0
stripe==5.10.0
bcrypt==4.1.1
passlib==1.7.4
pydantic==2.0.0
email-validator==2.0.0


